Modify to pick up hornblower-style notifications and turn ferry location red.
